requirements.txt
fastapi
uvicorn
python-dotenv
resend>=0.6.0